% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglm_mapper.R
\name{biglm_mapper}
\alias{biglm_mapper}
\title{Map-reduce with \code{biglm}}
\usage{
biglm_mapper(file_list, formula, levels_union = list(), ...)
}
\arguments{
\item{file_list}{Character vector of data file names. Must be in \code{.rds} format.}

\item{formula}{Model formula.}

\item{levels_union}{Named list. Names should be variable names, values should be character vectors.}

\item{...}{Additional arguments to pass to \code{biglm}}
}
\value{
A fitted \code{biglm} object.
}
\description{
Map-reduce with \code{biglm}
}
\examples{
biglm_mapper(data_files, y ~ x + z)
}
