% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglm_mapper.R
\name{cluster_se}
\alias{cluster_se}
\title{Map-reduce clustered standard errors with \code{biglm}}
\usage{
cluster_se(file_list, fitted_model, ...)
}
\arguments{
\item{file_list}{Character vector of data file names. Must be in \code{.rds} format.}

\item{fitted_model}{Fitted \code{biglm} object.}

\item{...}{Additional arguments to pass to \code{chunk_clusterse}
\itemize{
\item \code{clustervar}: character name of variable to cluster on
\item \code{levels_union}: named list with factor variables to standardize
}}
}
\value{
A list with elements
\itemize{
\item \code{xxinv}: (X'X)^-1
\item \code{cluster_vcov}: cluster-robust variance-covariance matrix estimate
\item \code{g_cov}: inner matrix in cluster-robust vcov matrix
\item \code{e_hat}: mean of squared residuals
}
}
\description{
Map-reduce clustered standard errors with \code{biglm}
}
\details{
Computation DOES NOT do small sample DOF adjustment (I assume if you're using \code{biglm}, the data is large).
Also importantly, computation assumes no group overlap across files (all observations from each group are in one file).
}
\examples{
cluster_se(data_files, m, clustervar="group")
}
