% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglm_mapper.R
\name{unite_levels}
\alias{unite_levels}
\title{Get list, one element per variable, of union of levels of a set of variables in a set of files.}
\usage{
unite_levels(file_list, factor_vars)
}
\arguments{
\item{file_list}{List of file names, in .rds format}

\item{factor_vars}{Character vector of variable names to find union of levels for}
}
\value{
Named list containing the union of values appearing in each variable in \code{factor_vars} across all files in \code{file_list}.
}
\description{
Get list, one element per variable, of union of levels of a set of variables in a set of files.
}
\examples{
unite_levels(files, c("factor_variable_1","factor_variable_2"))
}
