% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglm_mapper.R
\name{update_levels}
\alias{update_levels}
\title{Standardize levels of factor variables.}
\usage{
update_levels(data, levels_union)
}
\arguments{
\item{data}{Dataset to standardize}

\item{levels_union}{Named list. Names should be variable names, values should be character vectors.}
}
\value{
\code{data} with variables in \code{levels_union} converted to factors with levels specified in \code{levels_union}
}
\description{
Standardize levels of factor variables.
}
\examples{
update_levels(d, list(var1=c("a","b","c"), var2=c("x","y")))
}
